% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pipe.R
\name{\%then\%}
\alias{\%then\%}
\alias{\%catch\%}
\alias{\%finally\%}
\alias{\%then_tee\%}
\alias{\%catch_tee\%}
\alias{\%all\%}
\alias{\%race\%}
\title{Named promise pipe operators}
\usage{
lhs \%then\% rhs

lhs \%catch\% rhs

lhs \%finally\% rhs

lhs \%then_tee\% rhs

lhs \%catch_tee\% rhs

lhs \%all\% rhs

lhs \%race\% rhs
}
\arguments{
\item{lhs}{A promise object.}

\item{rhs}{A function call using the magrittr semantics. It can return either
a promise or non-promise value, or throw an error.}
}
\value{
A new promise.
}
\description{
Promise-aware pipe operators containing the method within the pipeline.
Like R's pipe operator,\verb{|>}, these operators can be used to chain together pipelines
of promise-transforming operations. Unlike \verb{|>}, these pipe operators methods wait
for promise resolution and pass the unwrapped value (or error) to the \code{rhs}
function call.
}
\details{
\verb{\%then\%} is for handling a successful resolution, while \verb{\%catch\%} is for
handling errors, and finally, \verb{\%finally\%} is for handling code that runs
after the already chained (successful or failed) promises have been handled.

The \verb{_tee} variants of each return the \code{lhs} instead of the \code{rhs}, which is
useful for pipeline steps that are used for side effects (printing, plotting,
saving). One situation where the \verb{_tee} variants break down is when \code{func()}
throws an error, or returns a promise that ultimately fails. In that case,
the failure will be propagated by our pipe operators but not by the
magrittr-plus-function \emph{equivalents}.

For simplicity of implementation, we do not support the \pkg{magrittr} feature of
using a \code{.} at the head of a pipeline to turn the entire pipeline into a
function instead of an expression.
}
\section{Functions}{
\itemize{
\item \code{lhs \%then\% rhs}: Handling a successful resolution. Equivalent to \code{promise |> then(func)}

\item \code{lhs \%catch\% rhs}: Handling an error. Equivalent to \code{promise |> catch(func)}

\item \code{lhs \%finally\% rhs}: Handling code that runs after already chained (successful or failed) promises have been handled. Equivalent to \code{promise |> finally(func)}

\item \code{lhs \%then_tee\% rhs}: Handling a successful resolution, returning the \code{lhs}. Equivalent to \code{promise }\%T>\%\code{ then(func, tee = TRUE)}, which uses the \pkg{magrittr} tee operator, \code{\link{\%T>\%}}.

\item \code{lhs \%catch_tee\% rhs}: Handling an error, returning the \code{lhs}. Equivalent to \code{promise |> catch(func, tee = TRUE)} or \code{promise \%T>\% catch(func)}, which uses the \pkg{magrittr} tee operator, \code{\link{\%T>\%}}.

\item \code{lhs \%all\% rhs}: Waits for all promises within the \code{lhs} list before passing through a list of results to the \code{rhs}. Equivalent to \code{promise_all(.list = lhs) \%then\% rhs}.

\item \code{lhs \%race\% rhs}: Waits for the first promise to resolve within the \code{lhs} list of promises before passing through the value from the promise that resolved first. Equivalent to \code{promise_race(.list = lhs) \%then\% rhs}.

}}
\examples{
\dontrun{
library(mirai)

mirai(cars) \%then\%
  head(5) \%then_tee\%
  print()

# If the read.csv fails, resolve to NULL instead
mirai(read.csv("http://example.com/data.csv")) \%catch\%
  { NULL }

# Return a list of results from a list of promises:
proms <- list(
  mirai({ Sys.sleep(1); 1 }),
  mirai({ Sys.sleep(0.5); 2 })
)
proms \%all\% print()
#> [[1]]
#> [1] 1
#>
#> [[2]]
#> [1] 2

# Return the first promise to resolve from a list of promises:
proms <- list(
  mirai({ Sys.sleep(1); 1 }),
  mirai({ Sys.sleep(0.5); 2 })
)
proms \%race\% print()
#> [1] 2

}

}
\seealso{
https://rstudio.github.io/promises/articles/promises_03_overview.html#using-pipes
}
