% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/otel-promise.R
\name{create_otel_active_span_promise_domain}
\alias{create_otel_active_span_promise_domain}
\alias{local_promise_domain}
\title{Create and manage otel promise domain / spans}
\usage{
create_otel_active_span_promise_domain(active_span)

local_promise_domain(
  domain,
  ...,
  replace = FALSE,
  local_envir = parent.frame()
)
}
\arguments{
\item{active_span}{An otel active span object. This span will must be
previously created and must be closed manually.}

\item{domain}{A promise domain object.}
}
\description{
Create and manage otel promise domain / spans
}
\section{Functions}{
\itemize{
\item \code{create_otel_active_span_promise_domain()}: Creates a promise domain for an active OpenTelemetry span.

\item \code{local_promise_domain()}: Local promise domain for an active OpenTelemetry
span. The promise domain will be removed when the \code{.envir} is removed from
the stack. This method is currently only compatible with
\code{create_otel_active_span_promise_domain(active_span)} as we can garantee
that the domain's \code{wrapSync} is \code{base::force}.

}}
\seealso{
\itemize{
\item \code{\link[=local_otel_active_span_promise_domain]{local_otel_active_span_promise_domain()}}
\item \code{\link[=with_otel_active_span_promise_domain]{with_otel_active_span_promise_domain()}}
}
}
\keyword{internal}
