% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/otel-promise.R
\name{otel_create_inactive_span}
\alias{otel_create_inactive_span}
\alias{otel_start_active_span}
\alias{local_otel_active_span_promise_domain}
\alias{with_otel_active_span_promise_domain}
\title{Manage promise domains given an active OpenTelemetry span.}
\usage{
otel_create_inactive_span(
  name,
  attributes = list(),
  links = NULL,
  option = NULL,
  ...
)

otel_start_active_span(
  name,
  ...,
  attributes = list(),
  activation_scope = parent.frame(),
  end_on_exit = FALSE
)

local_otel_active_span_promise_domain(
  active_span,
  ...,
  replace = FALSE,
  .envir = parent.frame()
)

with_otel_active_span_promise_domain(active_span, expr, ..., replace = FALSE)
}
\arguments{
\item{name}{The name of the span.}

\item{attributes}{A list of attributes to add to the span. This value will be passed into \code{otel::as_attributes()}.}

\item{links}{A list of links to add to the span. Default is \code{NULL}.}

\item{option}{Additional options for the span. Default is \code{NULL}.}

\item{...}{Additional arguments (not used).}

\item{activation_scope}{The environment in which the span will be activated.}

\item{end_on_exit}{If \code{TRUE}, the span will be ended when the activation scope is exited. Default is \code{FALSE}. When \code{FALSE}, the span must be ended manually using \code{otel::end_span(spn)}.}

\item{active_span}{An active span object.}

\item{replace}{If \code{TRUE}, replaces the current promise domain with the new one.}

\item{.envir}{The environment in which to create the local promise domain.}

\item{expr}{An expression to evaluate within the promise domain.}
}
\description{
Manage promise domains given an active OpenTelemetry span.
}
\section{Functions}{
\itemize{
\item \code{otel_create_inactive_span()}: Create (but do not start) an OpenTelemetry span that will be used
within a promise domain. The span will need to be ended manually (\code{otel::end_span(spn)}).
The returned object is intended to be used with \code{otel::local_active_span(spn)} or
\code{otel::with_active_span(spn, expr)}.

\item \code{otel_start_active_span()}: Create and start an active OpenTelemetry span. This span will need to be ended manually (\code{otel::end_span(spn)}).

\item \code{local_otel_active_span_promise_domain()}: Local promise domain for an active OpenTelemetry span. The
promise domain will be removed when the \code{.envir} is removed from the stack.

\item \code{with_otel_active_span_promise_domain()}: Local promise domain for an active OpenTelemetry span. The
promise domain will be removed when the \code{.envir} is removed from the stack.

}}
\examples{
\dontrun{
# Example usage:
library(promises)
library(otel)

# Create an active span
inactive_span <- otel_create_inactive_span("example_span")
# Use the local promise domain
local_otel_active_span_promise_domain(inactive_span)

# Create a promise within the active span domain
p <- promise_resolve(TRUE) \%...>\%
  {
    # This will run within the active span domain
    # Run user code here
    Sys.sleep(1)
    "Done"
  } \%...>\%
  {
    # This will run within the active span domain
    print(.)
  }

p <- finally(p, function() {
  # This will run after the promise is resolved
  otel::end_span(inactive_span)
})
}
}
\seealso{
\itemize{
\item \code{\link[=create_otel_active_span_promise_domain]{create_otel_active_span_promise_domain()}}
\item \code{\link[=local_promise_domain]{local_promise_domain()}}
\item \code{\link[=with_promise_domain]{with_promise_domain()}}
}
}
\keyword{internal}
