% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/env.R
\name{promise_env}
\alias{promise_env}
\title{Environment-based Promise implementation}
\usage{
promise_env()
}
\value{
A promise object
}
\description{
This is an environment-based implementation of the Promise class,
converted from the original R6 implementation.
Create a new promise using native environments
}
