% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/env.R
\name{promise_obj}
\alias{promise_obj}
\title{Create a promise wrapper structure}
\usage{
promise_obj(promise_impl)
}
\arguments{
\item{promise_impl}{The environment-based promise implementation}
}
\value{
A promise object with the standard promise interface
}
\description{
Create a promise wrapper structure
}
