% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deprecated.R
\name{promises-deprecated}
\alias{promises-deprecated}
\alias{with_ospan_async}
\alias{with_ospan_promise_domain}
\alias{local_ospan_promise_domain}
\title{\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#deprecated}{\figure{lifecycle-deprecated.svg}{options: alt='[Deprecated]'}}}{\strong{[Deprecated]}}: Deprecated functions in promises}
\usage{
with_ospan_async(name, expr, ..., tracer, attributes = NULL)

with_ospan_promise_domain(expr)

local_ospan_promise_domain(envir = parent.frame())
}
\description{
These functions have been deprecated and will be removed in a future version.
Please use the suggested alternatives instead.
}
\section{Functions}{
\itemize{
\item \code{with_ospan_async()}: Use \code{with_otel_span()} instead

\item \code{with_ospan_promise_domain()}: Use \code{with_otel_promise_domain()} instead

\item \code{local_ospan_promise_domain()}: Use \code{local_otel_promise_domain()} instead

}}
\keyword{internal}
