% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/promises-package.R
\docType{package}
\name{promises-package}
\alias{promises}
\alias{promises-package}
\title{promises: Abstractions for Promise-Based Asynchronous Programming}
\description{
\if{html}{\figure{logo.svg}{options: style='float: right' alt='logo' width='120'}}

Provides fundamental abstractions for doing asynchronous programming in R using promises. Asynchronous programming is useful for allowing a single R process to orchestrate multiple tasks in the background while also attending to something else. Semantics are similar to 'JavaScript' promises, but with a syntax that is idiomatic R.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://rstudio.github.io/promises/}
  \item \url{https://github.com/rstudio/promises}
  \item Report bugs at \url{https://github.com/rstudio/promises/issues}
}

}
\author{
\strong{Maintainer}: Barret Schloerke \email{barret@posit.co} (\href{https://orcid.org/0000-0001-9986-114X}{ORCID})

Authors:
\itemize{
  \item Barret Schloerke \email{barret@posit.co} (\href{https://orcid.org/0000-0001-9986-114X}{ORCID})
  \item Joe Cheng \email{joe@posit.co}
  \item Winston Chang \email{winston@posit.co} (\href{https://orcid.org/0000-0002-1576-2126}{ORCID})
  \item Charlie Gao \email{charlie.gao@posit.co} (\href{https://orcid.org/0000-0002-0750-061X}{ORCID})
}

Other contributors:
\itemize{
  \item Posit Software, PBC (\href{https://ror.org/03wc8by49}{ROR}) [copyright holder, funder]
}

}
\keyword{internal}
