% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/promise.R
\name{resolve}
\alias{resolve}
\title{Fulfill a promise}
\usage{
resolve(value = NULL)
}
\arguments{
\item{value}{The result from a successful calculation.}

\item{reason}{An error or string that explains why the operation failed.}
}
\description{
Use these functions to satisfy a promise with either success (\code{resolve})
or failure (\code{reject}).
}
\keyword{internal}
