% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{wait_for}
\alias{wait_for}
\title{Wait for a Promise}
\usage{
wait_for(promise, loop = later::current_loop())
}
\arguments{
\item{promise}{A promise object. The object can be in any state.}

\item{loop}{the \pkg{later} loop on which the promise has been scheduled.
Defaults to \code{\link[later:create_loop]{later::current_loop()}}.}
}
\value{
The value of the resolved promise. An error will be thrown if the
promise is rejected.
}
\description{
Synchronizes a promise by waiting for it to complete and then retrieving its
value.
}
\examples{
# Returns a promise for the sum of e1 + e2, with a 1 sec delay
slowly_add <- function(e1, e2) {
  promise(\(resolve, reject) {
    later::later(\() resolve(e1 + e2), delay = 1)
  })
}

# Returns 3 after 1 second
res <- wait_for(slowly_add(1, 2))
res

}
