% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/promise.R
\name{is.promise}
\alias{is.promise}
\title{Tests whether an object is a promise}
\usage{
is.promise(x)
}
\arguments{
\item{x}{An R object to test.}
}
\description{
Use \code{is.promise} to determine whether an R object is a promise.
}
