% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/env.R
\name{promise}
\alias{promise}
\title{Constructor function that matches the original promise() interface}
\usage{
promise(action)
}
\arguments{
\item{action}{A function with signature \verb{function(resolve, reject)}}
}
\value{
A promise object
}
\description{
Constructor function that matches the original promise() interface
}
