% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{promise_all}
\alias{promise_all}
\alias{promise_race}
\alias{promise_lapply}
\title{Combine multiple promise objects}
\usage{
promise_all(..., list = NULL)

promise_race(..., list = NULL)

promise_lapply(X, FUN, ...)
}
\arguments{
\item{...}{Promise objects. Either all arguments must be named, or all
arguments must be unnamed. If \code{list} is provided, then these arguments are
ignored.}

\item{list}{A list of promise objects--an alternative to \code{...}.}

\item{X}{A vector (atomic or list) or an expression object. Other objects
(including classed objects) will be coerced by base::as.list.}

\item{FUN}{The function to be applied to each element of \code{X}. The function is
permitted, but not required, to return a promise.}
}
\value{
A promise.

For \code{promise_all}, if all of the promises were successful, the returned
promise will resolve to a list of the promises' values; if any promise
fails, the first error to be encountered will be used to reject the
returned promise.

For \code{promise_race}, the first of the promises to either fulfill or reject
will be passed through to the returned promise.
}
\description{
Use \code{promise_all} to wait for multiple promise objects to all be successfully
fulfilled. Use \code{promise_race} to wait for the first of multiple promise
objects to be either fulfilled or rejected.
}
\examples{
# TODO

}
