% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/promises-package.R
\docType{package}
\name{promises-package}
\alias{promises}
\alias{promises-package}
\title{promises: Abstractions for Promise-Based Asynchronous Programming}
\description{
Provides fundamental abstractions for doing asynchronous programming in R using promises. Asynchronous programming is useful for allowing a single R process to orchestrate multiple tasks in the background while also attending to something else. Semantics are similar to 'JavaScript' promises, but with a syntax that is idiomatic R.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://rstudio.github.io/promises/}
  \item \url{https://github.com/rstudio/promises}
  \item Report bugs at \url{https://github.com/rstudio/promises/issues}
}

}
\author{
\strong{Maintainer}: Joe Cheng \email{joe@posit.co}

Other contributors:
\itemize{
  \item Posit Software, PBC [copyright holder, funder]
}

}
\keyword{internal}
